Clean up naming and comments and remove intermediate files
